backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
pydantic[email]==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
